+++ packages/core/src/SDKConfig/SDKConfig.ts
@@ -79,7 +79,12 @@ export interface SDKConfig {
   accessTokenExpireCookieName?: string;
 
   /**
+   * Callback to be invoked if a request to refresh the access token fails during autorefresh.
+   */
+  onAutoRefreshFailure?: (error: Error) => void;
+
+  /**
    * Callback to be invoked at the moment of access token expiration
    */
-  onTokenExpiration?: () => void;
+  onTokenExpiration: () => void;
 }