+++ packages/sdk-react/src/components/providers/FusionAuthProviderConfig.ts
@@ -43,6 +43,11 @@ export interface FusionAuthProviderConfig {
   onRedirect?: (state?: string) => void;
 
   /**
+   * Callback to be invoked if a request to refresh the access token fails during autorefresh.
+   */
+  onAutoRefreshFailure?: (error: Error) => void;
+
+  /**
    * The path to the login endpoint.
    */
   loginPath?: string;